+++ tsconfig.json
@@ -24,6 +24,6 @@
     "typeRoots": ["./src/types", "node_modules/@types"]
   },
   "files": ["./src/types/svg.d.ts", "src/types/global.d.ts"],
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "inapp/**/*"],
   "exclude": ["node_modules/**/*"]
 }
